Create TracksContainer to handle track scroll
This way the tracks can be scrolled sideways with the mouse wheel even
when on the track resize handles or scale bar.
